+++ packages/cli/test/dev/integration-2.test.ts
@@ -200,7 +200,6 @@ test(
 );
 
 // Skipping because it doesn't run yet on Node 22
-// eslint-disable-next-line jest/no-disabled-tests
 test.skip(
   '[vercel dev] Should set the `ts-node` "target" to match Node.js version',
   testFixtureStdio('node-ts-node-target', async (testPath: any) => {