+++ packages/node/test/integration/integration-setup.js
@@ -25,11 +25,9 @@ module.exports = function setupTests(groupIndex) {
     console.log('testing group', groupIndex, fixtures);
   }
 
-  // eslint-disable-next-line no-restricted-syntax
   for (const fixture of fixtures) {
     const errMsg = testsThatFailToBuild.get(fixture);
     if (errMsg) {
-      // eslint-disable-next-line no-loop-func
       it(`should fail to build ${fixture}`, async () => {
         try {
           await testDeployment(path.join(fixturesPath, fixture));
@@ -39,7 +37,7 @@ module.exports = function setupTests(groupIndex) {
           expect(err.deployment.errorMessage).toBe(errMsg);
         }
       });
-      continue; //eslint-disable-line
+      continue;
     }
     it(`should build ${fixture}`, async () => {
       await expect(