+++ packages/hono/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/hono
 
+## 0.2.47
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.6.11
+
+## 0.2.46
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.6.10
+
 ## 0.2.45
 
 ### Patch Changes